refactor(adapter): consolidate source/adapter architecture with pluggable registry (#193)
- Extract Dialect behaviour from old Lotus.Source callbacks
- Move dialect modules to Lotus.Source.Adapters.Ecto.Dialects.*
- Transform Lotus.Source from behaviour into facade
- Consolidate three duplicated @impls maps into one
- Add can_handle?/1 and wrap/2 optional adapter callbacks
- Add source_adapters config for external adapter registration
- Add use Lotus.Source.Adapters.Ecto macro for external Ecto adapters
- Update Lotus.Source.Resolvers.Static for pluggable resolution
- Migrate TypeMapper to source_type atoms (:postgres, :mysql, :sqlite)
- Migrate all callers from Lotus.Sources to Lotus.Source
- Old Lotus.Sources.* modules delegate for backward compatibility
