+++ lib/lotus/sources.ex
@@ -1,165 +1,18 @@
 defmodule Lotus.Sources do
   @moduledoc false
 
-  alias Lotus.Config
-  alias Lotus.Source.Adapter
+  # Deprecated: use Lotus.Source instead.
+  # This module delegates all calls to the new facade for backward compatibility.
 
-  @doc """
-  Resolve to an `%Adapter{}` struct.
-
-  Accepts:
-    * `source_opt` — configured name (string) or source module (atom) or nil
-    * `q_source`   — query's stored source (string or module) or nil
-
-  Falls back to the default source. Raises on resolution failure.
-  """
-  @spec resolve!(nil | String.t() | module(), nil | String.t() | module()) :: Adapter.t()
-  def resolve!(source_opt, q_source) do
-    case resolver().resolve(source_opt, q_source) do
-      {:ok, %Adapter{} = adapter} ->
-        adapter
-
-      {:error, :not_found} ->
-        available = resolver().list_source_names()
-        label = source_opt || q_source
-
-        raise ArgumentError,
-              "Data source '#{label}' not configured. " <>
-                "Available sources: #{inspect(available)}"
-    end
-  end
-
-  @doc """
-  Lists all configured source adapters.
-  """
-  @spec list_sources() :: [Adapter.t()]
-  def list_sources, do: resolver().list_sources()
-
-  @doc """
-  Gets a source adapter by name. Raises if not found.
-  """
-  @spec get_source!(String.t()) :: Adapter.t()
-  def get_source!(name), do: resolver().get_source!(name)
-
-  @doc """
-  Returns the default source as an `%Adapter{}` struct.
-  """
-  @spec default_source() :: Adapter.t()
-  def default_source do
-    {_name, adapter} = resolver().default_source()
-    adapter
-  end
-
-  @doc false
-  @spec name_from_module!(module()) :: String.t()
-  def name_from_module!(mod) do
-    case Enum.find(Config.data_sources(), fn {_name, m} -> m == mod end) do
-      {name, _} ->
-        name
-
-      nil ->
-        raise ArgumentError,
-              "Source module #{inspect(mod)} isn't in :lotus, :data_sources. " <>
-                "Configured names: #{inspect(Map.keys(Config.data_sources()))}"
-    end
-  end
-
-  @doc """
-  Detect the source type from an adapter, repository module, or name.
-  """
-  @spec source_type(Adapter.t() | module() | String.t()) ::
-          :postgres | :mysql | :sqlite | :other
-  def source_type(%Adapter{source_type: st}), do: st
-
-  def source_type(repo_name) when is_binary(repo_name) do
-    repo = Config.get_data_source!(repo_name)
-    source_type(repo)
-  end
-
-  def source_type(repo) when is_atom(repo) do
-    case repo.__adapter__() do
-      Ecto.Adapters.Postgres -> :postgres
-      Ecto.Adapters.SQLite3 -> :sqlite
-      Ecto.Adapters.MyXQL -> :mysql
-      _ -> :other
-    end
-  end
-
-  @doc """
-  Whether a source supports a specific feature.
-
-  Accepts an `%Adapter{}` struct (delegates to the adapter callback) or
-  a source type atom (uses hardcoded fallback for backward compatibility).
-  """
-  @spec supports_feature?(Adapter.t() | atom(), atom()) :: boolean()
-  def supports_feature?(%Adapter{} = adapter, feature) do
-    Adapter.supports_feature?(adapter, feature)
-  end
-
-  def supports_feature?(:postgres, :schema_hierarchy), do: true
-
-  def supports_feature?(:postgres, :search_path), do: true
-  def supports_feature?(:postgres, :make_interval), do: true
-  def supports_feature?(:postgres, :arrays), do: true
-  def supports_feature?(:postgres, :json), do: true
-
-  def supports_feature?(:mysql, :schema_hierarchy), do: false
-
-  def supports_feature?(:mysql, :search_path), do: false
-  def supports_feature?(:mysql, :make_interval), do: false
-  def supports_feature?(:mysql, :arrays), do: false
-  def supports_feature?(:mysql, :json), do: true
-
-  def supports_feature?(:sqlite, :schema_hierarchy), do: false
-
-  def supports_feature?(:sqlite, :search_path), do: false
-  def supports_feature?(:sqlite, :make_interval), do: false
-  def supports_feature?(:sqlite, :arrays), do: false
-  def supports_feature?(:sqlite, :json), do: true
-
-  def supports_feature?(_, _), do: false
-
-  @doc """
-  Return the human-readable label for the top-level hierarchy in a source.
-  """
-  @spec hierarchy_label(Adapter.t() | String.t()) :: String.t()
-  def hierarchy_label(%Adapter{} = adapter), do: Adapter.hierarchy_label(adapter)
-
-  def hierarchy_label(source_name) when is_binary(source_name) do
-    source_name |> get_source!() |> Adapter.hierarchy_label()
-  end
-
-  @doc """
-  Return an example query string suitable for placeholder text.
-  """
-  @spec example_query(Adapter.t() | String.t(), String.t(), String.t() | nil) :: String.t()
-  def example_query(%Adapter{} = adapter, table, schema),
-    do: Adapter.example_query(adapter, table, schema)
-
-  def example_query(source_name, table, schema) when is_binary(source_name) do
-    source_name |> get_source!() |> Adapter.example_query(table, schema)
-  end
-
-  @doc """
-  Return the query language identifier for a source.
-  """
-  @spec query_language(Adapter.t() | String.t()) :: String.t()
-  def query_language(%Adapter{} = adapter), do: Adapter.query_language(adapter)
-
-  def query_language(source_name) when is_binary(source_name) do
-    source_name |> get_source!() |> Adapter.query_language()
-  end
-
-  @doc """
-  Wrap a statement with a limit clause using the source's syntax.
-  """
-  @spec limit_query(Adapter.t() | String.t(), String.t(), pos_integer()) :: String.t()
-  def limit_query(%Adapter{} = adapter, statement, limit),
-    do: Adapter.limit_query(adapter, statement, limit)
-
-  def limit_query(source_name, statement, limit) when is_binary(source_name) do
-    source_name |> get_source!() |> Adapter.limit_query(statement, limit)
-  end
-
-  defp resolver, do: Config.source_resolver()
+  defdelegate resolve!(source_opt, q_source), to: Lotus.Source
+  defdelegate list_sources(), to: Lotus.Source
+  defdelegate get_source!(name), to: Lotus.Source
+  defdelegate default_source(), to: Lotus.Source
+  defdelegate name_from_module!(mod), to: Lotus.Source
+  defdelegate source_type(source), to: Lotus.Source
+  defdelegate supports_feature?(source, feature), to: Lotus.Source
+  defdelegate hierarchy_label(source), to: Lotus.Source
+  defdelegate example_query(source, table, schema), to: Lotus.Source
+  defdelegate query_language(source), to: Lotus.Source
+  defdelegate limit_query(source, statement, limit), to: Lotus.Source
 end
