+++ lib/lotus/sources/default.ex
@@ -1,188 +1,31 @@
 defmodule Lotus.Sources.Default do
-  @moduledoc """
-  Default source adapter implementation for unsupported or unknown database sources.
+  @moduledoc false
 
-  Provides safe no-op implementations for source-specific functions and
-  generic error formatting for database errors.
-  """
+  # Deprecated: use Lotus.Source.Adapters.Ecto.Dialects.Default instead.
 
-  @behaviour Lotus.Source
+  alias Lotus.Source.Adapters.Ecto.Dialects.Default, as: DefaultDialect
 
-  alias Lotus.SQL.FilterInjector
-  alias Lotus.SQL.SortInjector
-
-  @impl true
-  @doc "Simple transaction wrapper for unsupported sources."
-  def execute_in_transaction(repo, fun, opts) do
-    timeout = Keyword.get(opts, :timeout, 15_000)
-
-    repo.transaction(fun, timeout: timeout)
-  rescue
-    e -> {:error, Exception.message(e)}
-  end
-
-  @impl true
-  @doc "No-op: unsupported sources do not implement statement timeouts."
-  def set_statement_timeout(_repo, _ms), do: :ok
-
-  @impl true
-  @doc "No-op: unsupported sources do not implement search_path."
-  def set_search_path(_repo, _path), do: :ok
-
-  @impl true
-  @doc """
-  Formats common error types into strings. Falls back to `inspect/1`
-  for unknown values.
-  """
-  def format_error(%{__exception__: true} = e), do: Exception.message(e)
-  def format_error(%DBConnection.EncodeError{message: msg}), do: msg
-  def format_error(%ArgumentError{message: msg}), do: msg
-  def format_error(msg) when is_binary(msg), do: msg
-  def format_error(other), do: "Database Error: #{inspect(other)}"
-
-  @impl true
-  @doc """
-  Returns a generic SQL parameter placeholder (`"?"`).
-
-  This keeps the query builder working even for unknown sources,
-  though actual binding semantics may differ.
-  """
-  def param_placeholder(_idx, _var, _type), do: "?"
-
-  @impl true
-  @doc """
-  Returns generic placeholders for LIMIT and OFFSET (`"?"` for both).
-
-  Unknown sources should implement their own version if they have specific requirements.
-  """
-  def limit_offset_placeholders(_limit_idx, _offset_idx) do
-    {"?", "?"}
-  end
-
-  @impl true
-  @doc "The default source does not handle any specific exceptions."
-  def handled_errors, do: []
-
-  @impl true
-  def query_language, do: "sql"
-
-  @impl true
-  def limit_query(statement, limit) do
-    "SELECT * FROM (#{statement}) AS limited_query LIMIT #{limit}"
-  end
-
-  @impl true
-  @doc """
-  Returns conservative deny rules covering common system tables from various databases.
-
-  Since we don't know the specific source, we include deny rules for all known
-  system tables to be safe.
-  """
-  def builtin_denies(_repo) do
-    [
-      {"pg_catalog", ~r/.*/},
-      {"information_schema", ~r/.*/},
-      {nil, ~r/^sqlite_/},
-      {nil, "schema_migrations"},
-      {"public", "schema_migrations"},
-      {"public", "lotus_queries"},
-      {nil, "lotus_queries"},
-      {"public", "lotus_query_visualizations"},
-      {nil, "lotus_query_visualizations"},
-      {"public", "lotus_dashboards"},
-      {nil, "lotus_dashboards"},
-      {"public", "lotus_dashboard_cards"},
-      {nil, "lotus_dashboard_cards"},
-      {"public", "lotus_dashboard_filters"},
-      {nil, "lotus_dashboard_filters"},
-      {"public", "lotus_dashboard_card_filter_mappings"},
-      {nil, "lotus_dashboard_card_filter_mappings"}
-    ]
-  end
-
-  @impl true
-  def default_schemas(_repo) do
-    # Conservative default for unknown sources
-    ["public"]
-  end
-
-  @impl true
-  def supports_feature?(_), do: false
-
-  @impl true
-  def hierarchy_label, do: "Tables"
-
-  @impl true
-  def example_query(table, _schema) do
-    "SELECT value_column FROM #{table}"
-  end
-
-  @impl true
-  def builtin_schema_denies(_repo) do
-    # Conservative denies covering common system schemas from various databases
-    ["pg_catalog", "information_schema", "mysql", "performance_schema", "sys"]
-  end
-
-  @impl true
-  @doc """
-  Generic list_schemas implementation that returns an empty list.
-  Unknown sources should implement their own version if they support schema introspection.
-  """
-  def list_schemas(_repo) do
-    []
-  end
-
-  @impl true
-  @doc """
-  Generic list_tables implementation that returns an empty list.
-  Unknown sources should implement their own version if they support schema introspection.
-  """
-  def list_tables(_repo, _schemas, _include_views?) do
-    []
-  end
-
-  @impl true
-  @doc """
-  Generic get_table_schema implementation that returns an empty schema.
-  Unknown sources should implement their own version if they support schema introspection.
-  """
-  def get_table_schema(_repo, _schema, _table) do
-    []
-  end
-
-  @impl true
-  @doc """
-  Returns an error since the database type is unknown and EXPLAIN syntax varies.
-  """
-  def explain_plan(_repo, _sql, _params, _opts) do
-    {:error, "EXPLAIN not supported for this database adapter"}
-  end
-
-  @impl true
-  @doc """
-  Generic resolve_table_schema that always returns nil.
-  This is appropriate for databases without schema support or unknown sources.
-  """
-  def resolve_table_schema(_repo, _table, _schemas) do
-    nil
-  end
-
-  @impl true
-  @doc "Double-quotes identifiers as a safe default for unknown SQL sources."
-  def quote_identifier(identifier) do
-    escaped = String.replace(identifier, "\"", "\"\"")
-    ~s("#{escaped}")
-  end
-
-  @impl true
-  @doc "Applies filters using standard SQL CTE wrapping with double-quoted identifiers."
-  def apply_filters(sql, params, filters) do
-    FilterInjector.apply(sql, params, filters, &quote_identifier/1, fn _idx -> "?" end)
-  end
-
-  @impl true
-  @doc "Applies sorts by appending ORDER BY clause with double-quoted identifiers."
-  def apply_sorts(sql, sorts) do
-    SortInjector.apply(sql, sorts, &quote_identifier/1)
-  end
+  defdelegate execute_in_transaction(repo, fun, opts), to: DefaultDialect
+  defdelegate set_statement_timeout(repo, ms), to: DefaultDialect
+  defdelegate set_search_path(repo, path), to: DefaultDialect
+  defdelegate format_error(error), to: DefaultDialect
+  defdelegate param_placeholder(idx, var, type), to: DefaultDialect
+  defdelegate limit_offset_placeholders(limit_idx, offset_idx), to: DefaultDialect
+  defdelegate handled_errors(), to: DefaultDialect
+  defdelegate query_language(), to: DefaultDialect
+  defdelegate limit_query(statement, limit), to: DefaultDialect
+  defdelegate builtin_denies(repo), to: DefaultDialect
+  defdelegate default_schemas(repo), to: DefaultDialect
+  defdelegate supports_feature?(feature), to: DefaultDialect
+  defdelegate hierarchy_label(), to: DefaultDialect
+  defdelegate example_query(table, schema), to: DefaultDialect
+  defdelegate builtin_schema_denies(repo), to: DefaultDialect
+  defdelegate list_schemas(repo), to: DefaultDialect
+  defdelegate list_tables(repo, schemas, include_views?), to: DefaultDialect
+  defdelegate get_table_schema(repo, schema, table), to: DefaultDialect
+  defdelegate explain_plan(repo, sql, params, opts), to: DefaultDialect
+  defdelegate resolve_table_schema(repo, table, schemas), to: DefaultDialect
+  defdelegate quote_identifier(identifier), to: DefaultDialect
+  defdelegate apply_filters(sql, params, filters), to: DefaultDialect
+  defdelegate apply_sorts(sql, sorts), to: DefaultDialect
 end
