+++ test/lotus/ai/query_explainer_test.exs
@@ -9,7 +9,7 @@ defmodule Lotus.AI.QueryExplainerTest do
 
       Mimic.copy(Lotus.Config)
 
-      stub(Lotus.Sources, :source_type, fn _ -> :postgres end)
+      stub(Lotus.Source, :source_type, fn _ -> :postgres end)
       stub(Lotus.Config, :get_data_source!, fn "postgres" -> Lotus.Test.Repo end)
 
       :ok
