+++ test/lotus/ai_test.exs
@@ -50,7 +50,7 @@ defmodule Lotus.AITest do
       setup_mocks()
 
       # Mock schema introspection
-      stub(Lotus.Sources, :source_type, fn _ -> :postgres end)
+      stub(Lotus.Source, :source_type, fn _ -> :postgres end)
       stub(Lotus.Schema, :list_tables, fn _ -> {:ok, table_list()} end)
 
       set_ai_config(enabled: true, api_key: "sk-test")
@@ -188,7 +188,7 @@ defmodule Lotus.AITest do
     setup do
       setup_mocks()
 
-      stub(Lotus.Sources, :source_type, fn _ -> :postgres end)
+      stub(Lotus.Source, :source_type, fn _ -> :postgres end)
       stub(Lotus.Schema, :list_tables, fn _ -> {:ok, table_list()} end)
 
       set_ai_config(enabled: true, api_key: "sk-test")
