+++ test/support/ai_case.ex
@@ -33,7 +33,7 @@ defmodule Lotus.AICase do
       # Mock Lotus.Schema calls for schema introspection
       setup do
         Mimic.copy(Lotus.Schema)
-        Mimic.copy(Lotus.Sources)
+        Mimic.copy(Lotus.Source)
         Mimic.copy(Lotus.SQL.Validator)
 
         stub(Lotus.SQL.Validator, :validate, fn _sql, _ds ->
